Refactor configure index test (#50)
## Problem

After recent changes in integration test to use the existing index if it
exists, the configure index test was required to be updated to consume
less resources.

## Solution

1. Refactored checkIfIndexExists code: 
    a. To support both data and control plane integration tests
b. Check if any index matches the criteria compared to old approach of
checking the first index of the list and otherwise create a new index.
2. For configure test: 
    a. Updated the beforeAll logic to check if an index exist 
b. Deleted the afterEach block to delete each index and save them for
future test



## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [X] None of the above: (explain here)
Added Integration tests

## Test Plan

Ran integration tests locally.