Fix path for gcp-starter env, add assert with retry mechanism, and add deprecation warning in vector_service.proto (#57)
## Problem

When running tests with gcp-starter env. in java sdk, there were a
couple of issues that needed to be addressed. In particular:
1. Create and listIndexes wasn't working.
2. Queries parameter was deprecated but the proto file wasn't updated
with a warning.
3. After upserting vectors into an index of gcp-starter env., there was
some latency in querying the vectors which caused tests to fail.

## Solution
This PR addresses the following:
1. Fixed path for gcp-starter env to address the createIndex and
listIndexes issue in gcp-starter env.
2. Add deprecation tag for queries parameter of VectorRequest in
vector_service.proto file.
3. Added assert with retry mechanism function to make sure the tests are
passing even though they fail for first couple of times. Given
gcp-starter is a free-tier and not as performant as pod indexes, the
retry mechanism will help ensure passing of the tests.

## Type of Change

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Ran integration tests suite.