Update index operations (#62)
## Problem

After adding open api generated code for control plane, there's a need
for a wrapper for the index operations.

## Solution

This PR adds a wrapper for index operations and associated integration
tests. Couple of the integration tests are currently commented out and
will be handled in the upcoming PR's.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Ran integration tests.