Add collections operations to `PineconeControlPlaneClient` with integration tests (#65)
## Problem
The Java SDK is currently missing collections functionality. Now that
we've added generated code for the OpenAPI spec, we can hook up
collections operations in `PineconeControlPlaneClient`.

We'd also like to add integration tests to cover collections.

## Solution
- Update `PineconeControlPlaneClient` to include collections operations
for create, list, delete, and describe.
- Add new integration test files: `CollectionTest` and
`CollectionErrorTest`.
- Add new helpers to `IndexManager` for creating and connecting to an
index by `indexName`, creating a collection and waiting for it to be
ready, and polling until an index is ready.
- Add new helper to `BuildUpsertRequest` for generating vectors by
dimension.

Bonuses:

- I added some logging configs in `build.gradle` for `test` and the
`integrationTest` task. This was primarily to help me debug things in CI
and log a bit more info to the console. We can tweak as needed, but I
think having something like this will be really helpful as opposed to
the generic output.
- Set `max-parallel: 1` in the `integration-test` job in the `pr.yml`
workflow. Because of the way `findIndexWithDimensionAndType` works we
cannot run integration tests in parallel without flakiness.
- Fixed an issue in `AssertRetry`. We were incrementing `delay` at the
class level which is a static variable, meaning if you re-used the
function in one place it would continue increasing the delay for all
subsequent calls based on the `backOff`.

## Type of Change
- [X] New feature (non-breaking change which adds functionality)

## Test Plan
Run integration tests to make sure `CollectionTest` and
`CollectionErrorTest` pass as expected. Integration tests should also be
passing in CI barring any flakiness.