Add data plane wrapper and accept sparse indices as unsigned 32-bit integer (#70)
## Problem

1. The SDK is missing a wrapper for data plane operations that helps
improve user interaction.
2. The sparse indices in Pinecone are `unsigned 32-bit integer` but
since Java doesn't offer the support for it natively, so there is a
datatype mismatch issue.

## Solution
1. Add a wrapper for data plane operations and broke down the input
request objects to java native datatypes except for Struct. The wrapper
contains operations for both blocking and the async stub. Also I have
added overloaded methods so its easier for users to call the data plane
operations.
Example:
`public UpsertResponse upsert(UpsertRequest request)`
vs. 
```
public UpsertResponse upsert(String id,
            List<Float> values,
            List<Long> sparseIndices,
            List<Float> sparseValues,
            com.google.protobuf.Struct metadata,
            String namespace)
```

2. The datatype for sparse indices in Pinecone db is `unsigned 32-bit
integer` so as a part of this PR, the sparse indices will be now
accepted as Java `long` instead of Java `int` with the input range of
`unsigned 32-bit integer` [0, 2^32 - 1] and everything outside of this
range will throw `PineconeValidationException`. As a part of this
change, I have added 3 classes relating to the Query Response in the
`unsigned_indices_model` package since we don't want to edit the java
generated classes:

- QueryResponseWithUnsignedIndices.java (reflects QueryResponse)
- ScoredVectorWithUnsignedIndices.java (reflects ScoredVector)
- SparseValuesWithUnsignedIndices.java (reflects SparseValues)

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Added unit and integration tests.