Refactor configs and disable collections and configure tests (#77)
## Problem

Currently, the Java SDK has PineconeClientConfig and
PineconeConnectionConfig which are required to be instantiated before a
user can call any dataplane operations. Although, recently we did create
a constructor that allowed users to call any data plane operations
without having the need of instantiating both configs. But in order to
improve the user-experience and align it with other SDK's, we would like
to combine the two config classes.

Secondly, the CollectionsTest and ConfigureIndexTest are failing
consistently causing inconvenience in CI pipeline.

## Solution

This PR removes PineconeClientConfig, PineconeConnectionConfig, and
PineconeClient classes, and instead introduces PineconeConfig which has
a required field of apiKey along with optional members such as host,
usage-context, and customManagedChannel, where the customManagedChannel
will allow the user to input custom gRPC channel for data plane
operations.
As a part of this PR, I have disabled ConfigureIndexTest and
CollectionsTest to improve testing pipeline.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Updated integration tests.