Add `user-agent` to control plane operations, add `sourceTag`, re-enable `Pinecone` client unit tests (#84)
## Problem
We're currently not sending `user-agent` along with HTTP calls for
control plane operations. We also want to rename `integrationId` to
`sourceTag` to conform with the new attribution paradigm.

## Solution

While working on these changes I felt like it made sense to expand the
number of constructors `Pinecone` supports to allow for easier passing
of `apiKey`, `PineconeConfig`, `sourceTag`, and `OkHttpClient`. See the
comments on the PR for details.

I also wanted to get the unit tests for the control plane client running
again, so I refactored the old commented file.

- Add `user-agent` to HTTP control plane operations using
`ApiClient.setUserAgent(PineconeConfig.getUserAgent())`. We already
apply the `user-agent` to the gRPC `metadata`:
https://github.com/pinecone-io/pinecone-java-client/blob/4581d5ef8c4db9709c7287f9cdf6379cb6d63873/src/main/java/io/pinecone/configs/PineconeConnection.java#L140.
- Rename `integrationId` -> `sourceTag`. Thread `integrationId` from
`Pinecone` -> `PineconeConfig`. Apply normalization rules for the
`sourceTag` slug:
  - Lower case
  - Condense multiple spaces to one space
  - Replace spaces with "_"
  - Filter to charset [a-z0-9_]
- Refactor `PineconeIndexOperationClientTest` to `PineconeTest`. Add new
JSON fixtures under `/test/java/resources/` to support the new tests.

## Type of Change
- [X] New feature (non-breaking change which adds functionality)

## Test Plan
You should be able to set the `PINECONE_DEBUG` environment variable to
see log output while running integration tests locally. This is the
approach I used for validating `source_tag` was added to the user-agent,
but it's a bit of setup to get it up and running locally.

I'm going to look at updating some of the integration tests to
explicitly pass a test `sourceTag` for testing and we could validate
those results in Datadog.