Update OkHttpClient dependency version to 4.12.0 to address vulnerability issues and clean up codebase (#86)
## Problem

1. Address vulnerability concerns for OkHttpClient
2. Remove unnecessary logging
3. Data plane tests were failing lately
4. Run CI integration tests in parallel
5. Cleanup codebase

## Solution

As a part of this PR, I have updated the OkHttpClient version to 4.12.0
to address the vulnerability concerns. Secondly, I have removed --info
tag from integrationTest runs in CI because we don't need excessive
logging. Next, I have cleaned up the code base and removed commented
code from test files, and updated the data plane tests by adding a delay
of 2 mins (120000 ms) for allowing vectors to upsert for serverless
index test. Lastly, I have removed the constraint of running two tests
in sequence. The configure index test is failing for now but
@austin-denoble has an upcoming PR.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Ran integration tests locally and in CI.