Poll for Index ready during cleanup in `ConfigureIndexTest` (#87)
## Problem
We've started seeing consistent failures with the `ConfigureIndexTest`. 

The issue we're seeing is we're trying to delete the index before it has
finished upgrading:
![Screenshot 2024-03-27 at 11 45
30 AM](https://github.com/pinecone-io/pinecone-java-client/assets/119623786/405c85b6-a58e-4751-827d-b9f29de43c02)

```
io.pinecone.exceptions.PineconeBadRequestException: {"error":{"code":"INVALID_ARGUMENT","message":"Bad request: Cannot delete an index that is upgrading"},"status":400}
	at io.pinecone.exceptions.HttpErrorMapper.mapHttpStatusError(HttpErrorMapper.java:9)
	at io.pinecone.clients.Pinecone.handleApiException(Pinecone.java:167)
	at io.pinecone.clients.Pinecone.deleteIndex(Pinecone.java:111)
	at io.pinecone.integration.controlPlane.pod.ConfigureIndexTest.cleanUp(ConfigureIndexTest.java:34)
	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:64)
	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
	at java.base/java.lang.reflect.Method.invoke(Method.java:564)
	at org.junit.platform.commons.util.ReflectionUtils.invokeMethod(ReflectionUtils.java:725)
	at org.junit.jupiter.engine.execution.MethodInvocation.proceed(MethodInvocation.java:60)
        ...
```
Initially, I tried implementing polling using `waitUntilIndexIsReady()`
before each test and when cleaning up the index to make sure we're not
trying to operate on an `Upgrading` index. I noticed that there are
times when an index is marked as `ready:true` but `state=Initializing`
which seems to cause problems.

## Solution
I've added a `waitUntilIndexUpgraded` function to `ConfigureIndexTest`.
This ensures that we're waiting for the index to return to a `Ready`
state before sending additional requests or deleting.

![Screenshot 2024-03-27 at 12 03
34 PM](https://github.com/pinecone-io/pinecone-java-client/assets/119623786/0dc3356d-74aa-4679-a7e2-33f95c8a6e20)

## Type of Change
- [X] Infrastructure change (CI configs, testing)

## Test Plan

Describe specific steps for validating this change.