Handle nulls for some creation and configure methods and provide alternative method for creating collections (#91)
## Problem

Nulls:
We are not currently handling nulls gracefully when they are passed to
the following methods in the Pinecone class:
- CreateIndex
- ConfigureIndex
- CreateCollection

This PR captures them and raises a `PineconeValidationException` when
they're encountered.

`Request` objects:
- It's currently cumbersome for users to create `Request` objects each
time they have to call the above methods.

This PR overrides the `createCollection` method with more intuitive
params for the user. This abstracts away the `createCollectionRequest`
object w/native Java types, so that it matches the strategy we have for
similar patterns in the
[dataplane](https://github.com/pinecone-io/pinecone-java-client/blob/main/src/main/java/io/pinecone/clients/Index.java#L37-L68).
Now, a user can simply pass 2 `String`s to in order to create a
collection (the `collectionName` and the `sourceIndex`). Within this
method, the `createCollectionRequest` object is built.

If this PR is approved, I will follow up with a PR to do the same for
`CreateIndex` and `ConfigureIndex`. These need a bit more thought, so
I'll make them into separate PRs.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Unit tests and integration tests included.

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1206975505478879