Add `TestIndexResourcesManager` and `CleanupAllTestResourcesListener` (#89)
## Problem
Our current integration test setup is a bit difficult to reason about
and debug. When adding new integration tests, it can be hard to
understand the mechanics of how the current `IndexManager` creates and
manages Index resources. We also end up with a lot of lingering index
resources as indexes created via `createIndexIfNotExistsDataPlane`
aren't generally cleaned up as a part of the tests they're used in.

Ultimately, I think we should move away from
`createIndexIfNotExistsDataPlane` and `findIndexWithDimensionAndType`
entirely and implement a way for indexes to be explicitly shared across
tests by name.

## Solution
- Add new `TestIndexResourcesManager` singleton to handle creating,
sharing, and cleaning up index resources across an integration test run.
- Add new `CleanupAllTestResourcesListener` which implements Junit's
`TestExecutionListener` allowing us to run the singleton cleanup at the
end of the entire test run.
- Update `/integration/controlPlane/*` tests to use the new pattern.

My goal here was to have a centralized `TestIndexResourcesManager` class
that is used across all our integration tests. This class should handle
global setup and teardown before and after _all_ test classes run.
Essentially, I wanted this class to act as a generator and manager for
external resources.

I wanted to also have the `TestIndexResourcesManager` only create the
required resources as needed. Instead of trying to create and seed
multiple indexes before any tests run, I'm trying to do this only at the
time a test that's included in the run requests the resource.

The strategy I've used here is to leverage the Junit Platform Launcher
`TestExecutionListener` interface:
https://junit.org/junit5/docs/5.0.3/api/org/junit/platform/launcher/TestExecutionListener.html

We can create a new class `CleanupAllTestResourcesListener` that
implements `TestExceutionListener` and overrides
`testPlanExecutionFinished`. We then need to register this listener
which I've done by adding the
`/META-INF/services/org.junit.platform.launcher.TestExecutionListener`
file to the `integration` module. Check the code comments for a bit more
info about where I figured out how to do this.

In this PR I've only updated the tests under `/integration/controlPlane`
to leverage this new pattern, and it would be very similar for updating
the data plane operations. Basically, we would go through and replace
usages of `findIndexWithDimensionAndType`, `createNewIndex`, and
`createIndexIfNotExistsDataPlane` with calls to `getPodIndex()` from the
singleton.

I tried a few different methods of doing "global setup and teardown" in
Junit and found it quite difficult. It was a lot easier to trigger code
before and after each test class rather than the entirety of the run.

Other approaches tried:
- I tried implementing a Junit extension and using the `@ExtendWith()`
annotation across all tests:
https://junit.org/junit5/docs/current/user-guide/#extensions. This
worked for running code before and after each class but not the suite as
a whole.
- I tried having a base class `BaseTest` that was extended by each test
class, and had its own `@BeforeAll` and `@AfterAll`. Again, this
triggers every test class and not at the test run level.
- Using "test suites", or organizing our tests at a higher level than
discrete classes. This ultimately felt very cumbersome, and I couldn't
find straightforward ways of running a suite directly with gradle, or
handling running tests individually outside the suite context without
losing the setup and teardown.

 
## Type of Change
- [X] Infrastructure change (CI configs, etc)

## Test Plan
Run integration tests locally and in CI. Make sure control plane tests
pass as expected with this new architecture.