Add concurrent HashMap for storing indexNames and connection objects (#92)
## Problem

Everytime createIndexConnection() or createAsyncConnection() is called,
we are internally calling describeIndex() to get index host which is
required for creating blocking or async stubs.

## Solution

Create a singleton ConcurrentHashMap of indexName (key) and
PineconeConnection (value) such that every time the PineconeConnection
object is instantiated, the indexName along with the corresponding
PineconeConnection is stored into the HashMap. The entry is removed when
the connection object is closed.
Also for readability reasons, @ssmith-pc suggested to update the
`createIndexConnection()` and `createAsyncIndexConnection()` methods to
`getIndexConnection()` and `getAsyncConnection()` methods.

## Type of Change

- [X] New feature (non-breaking change which adds functionality)

## Test Plan

Added unit tests and ran integration tests.