Deprecate createIndex method (#105)
## Problem

Now that we have the new create index methods (`createPodsIndex` and
`createServerlessIndex`), we need to deprecate `createIndex` and update
the places in which it's used throughout the codebase.

Related ticket:
https://app.asana.com/0/1203260648987893/1206994598304493/f

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

All unit tests and integration tests are running successfully locally. 🤞
CI/CD behaves!

Will do a quick follow with a PR to update the README accordingly.