Add new `queryByVectorId` and `queryByVector` functions to `IndexInterface` (#106)
## Problem
Because of the lack of optional arguments, we have a lot of overloaded
functions in `IndexInterface` which are meant to make working with the
data plane easier for users.

Currently, we're missing some variations of `queryByVectorId` which need
to be added. We're also missing any functionality for `queryByVector`
behavior which is a commonly used operations by users.

## Solution
- Add new function overloads to `IndexInterface` to support
`queryByVectorId` minimal requests with namespace and/or `includeValues`
and `includeMetadata`.
- Duplicate our `queryByVectorId` functions into new `queryByVector`
functions.

All of these functions are going through the existing validation logic,
so it should be safe to add these in calling the base `query` function.
Unfortunately, there's not any unit tests for `Index` or `AsyncIndex`
and those will need to be added in a follow up.

This is also a lot of overloads, I'm sure there is a better pattern for
handling this from a UX perspective.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X]  New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI tests, or running an example app locally.