Add javadoc docstrings to `IndexInterface`, `Index`, `AsyncIndex` classes (#109)
## Problem
The Java client is lacking docstrings and we're working on adding them.

`IndexInterface`, `Index`, and `AsyncIndex` all need javadoc docstrings
with code examples.

## Solution
- Add docstrings to `IndexInterface` and the implementing classes:
`Index` and `AsyncIndex`. The code examples are documented in the
implementing classes and they inherit the root documentation from
`IndexInterface`.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Pull this branch down and run `gradle generateJavadoc` locally. Even
with the errors and warnings you should still get the `/docs` output in
the project folder. You can open this in your browser.

Make sure none of the errors or warnings have increased / we're not
seeing any outside of the generated code package:
![Screenshot 2024-04-10 at 2 46
13 AM](https://github.com/pinecone-io/pinecone-java-client/assets/119623786/6a5a8c90-a133-4805-aea4-7bdba9e8116b)