Update README and examples to reflect v1 changes (#107)
## Problem

We are close to launching v1! With it, our README needs a facelift. Here
is that facelift 💅.
We have also updated the SDK version, user-agent string, and the
examples package.

Asana tix: 
- Audrey: https://app.asana.com/0/1203260648987893/1207042384514994/f
- Rohan: https://app.asana.com/0/1203260648987893/1206745140276838/f

## Solution

Update README, user-agent string, and examples for v1 java SDK. We have
currently removed the UpsertsAndQueriesConcurrentExample example but in
upcoming PR's, we will like to add it back.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [x] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

---------

Co-authored-by: rohanshah18 <rohan.s@pinecone.io>
Co-authored-by: Dave Rigby <dave.r@pinecone.io>