Refactor data plane tests to use `TestResourcesManager`, clean up lengthy `Thead.sleep()` calls, general clean up (#99)
## Problem
The Java integration tests have traditionally been very flakey, and have
taken a long time to run. For example, if freshness is lagged or delayed
we can end up waiting nearly 30 minutes for a single integration test
run to finish due to `assertWithRetry` being used with a backoff of 4.

Previously, I added a new `TestIndexResourcesManager` class to handle
managing shared index resources across all tests over the
`IndexManager`:
https://github.com/pinecone-io/pinecone-java-client/pull/89. I initially
only implemented the manager in control plane tests. It also needs to be
added to data plane tests.

## Solution
- Rename `TestIndexResourcesManager` to `TestResourcesManager`.
- Rename `IndexManager` to `TestUtilities` and remove unneeded functions
for managing indexes.
- Clean up some of the duplicative function in `TestResourcesManager`.
Opt for the naming paradigm of `getOrCreateFoo` for the various
functions that allow getting or creating indexes or collections.
- Refactor all of the integration tests in `/integration/dataPlane/` to
use the `TestResourcesManager` rather than the helper functions in
`TestUtilities`.
- Remove all of the `Thread.sleep(90000)` etc calls that were sprinkled
all over the data plane tests. I don't think we need to actually wait
this long, it was incredibly inefficient and if freshness is degraded
they would still usually fail.
- For places where we're calling `assertWithRetry()` for retrying
commands in tests when things fail, I swapped from using a backoff of 4
to 3. We should look at refactoring this function to be a bit more
flexible, but for now we should avoid backoffs of > 3 as they cause an
enormous amount of wait time: 6000ms, 24000ms, 96000ms, 384000ms. We
should never wait this long within a test it would be preferable to let
things fail earlier and just re-run.
- I did a lot of various refactoring across the data plane tests.
Ultimately I don't think I've changed anything we're actually testing,
but I've tried to make things less brittle and more readable.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, integration tests etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Run the integration tests suite locally or validate we see green in CI.

`gradle integrationTest`