Add new `build-and-publish-docs.yml` GitHub Workflow (#116)
## Problem
The `pinecone-java-client` repo doesn't currently support building and
pushing client reference documentation to our docs repo like the
TypeScript and Python clients.

## Solution
Add new `build-and-publish-docs.yml` file to create a new workflow that
allows triggering a docs build and deployment.

For now I've left this as a workflow which can be manually dispatched,
or called from another workflow. We're reusing the `gradle
generateJavadoc` task which was added to the `build.gradle` file
previously.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
We'll need to trigger the workflow once it's in `main`.