Remove serverless public preview warning (#121)
## Problem
The serverless public preview warnings are still in the clients'
documentation from earlier in the year. Since we've expanded the number
of regions available, we should update these.

## Solution
Update the "Create a serverless index" example in the README to link to
the serverless documentation page.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
CI, run the client reference action after merged to main to update
reference docs.