Add ability to set PineconeClient managedChannel and Host
As is, customers have no way of setting custom managed channels or hosts with v1.0.0+ of pinecone java sdk, and this is a requirement if we want to tune the gRPC connection behaviour.

To that end, I have added two builder options:
1. withHost
2. withManagedChannel

In addition, I’ve made the package-private constructor for Pinecone class public so we can sidestep builder class if we ant.